middleware/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pyTigerGraph[gds]==1.8
apscheduler==3.10.4
pydantic==2.9.0
